feat: add safety limits and documentation for denied keys tracking
- Cap max_denied_keys at 10,000 to prevent excessive memory usage
- Add CLI validation (1-10000 range) with clear error messages
- Document the 3x growth factor memory behavior in code comments
- Add tests to verify the limit enforcement
- Update help text to show the valid range

+++ throttlecrab-server/src/main.rs
@@ -61,7 +61,7 @@ async fn main() -> Result<()> {
     // Create shared metrics instance
     let metrics = Arc::new(
         Metrics::builder()
-            .max_denied_keys(config.max_denied_keys)
+            .max_denied_keys(config.max_denied_keys as usize)
             .build(),
     );
